% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statUtil.R
\name{meanBeta}
\alias{meanBeta}
\title{Mean of the Beta distribution}
\usage{
meanBeta(alpha, beta)
}
\arguments{
\item{alpha}{Non-negative scalar}

\item{beta}{Non-negative scalar}
}
\description{
This function computes the mean of a Beta random variable
with parameters alpha and beta.
}
