% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDensity}
\alias{plotDensity}
\title{Plot a density function with ggplot}
\usage{
plotDensity(densFn, min, max, numPoints)
}
\arguments{
\item{densFn}{(Vectorised) function which returns density}

\item{min, max}{The min and max values for the x-axis}

\item{numPoints}{The number of points to plot}
}
\description{
Plot a density function with ggplot
}
