% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statUtil.R
\name{varBeta}
\alias{varBeta}
\title{Variance of the Beta distribution}
\usage{
varBeta(alpha, beta)
}
\arguments{
\item{alpha}{Non-negative scalar}

\item{beta}{Non-negative scalar}
}
\description{
This function computes the variance of a Beta random variable
with parameters alpha and beta.
}
